Conflicts:
	IGPE_Project/core/assets/rooms/r01_w-e.tmx
	IGPE_Project/core/assets/rooms/r02_w-e.tmx
	IGPE_Project/core/assets/rooms/r03_n-s.tmx
	IGPE_Project/core/assets/rooms/r04_n-s.tmx
	IGPE_Project/core/assets/rooms/r05_e-w.tmx
	IGPE_Project/core/assets/rooms/r06_n-e.tmx
	IGPE_Project/core/assets/rooms/r07_n-w.tmx
	IGPE_Project/core/assets/rooms/r08_w-n-e.tmx
	IGPE_Project/core/assets/rooms/r09_n-s-w.tmx
	IGPE_Project/core/assets/rooms/r10_n-s-e.tmx
	IGPE_Project/core/assets/rooms/r11_n-w-s-e.tmx
	IGPE_Project/core/assets/rooms/r12_s-e.tmx
	IGPE_Project/core/assets/rooms/rt1_w.tmx
	IGPE_Project/core/assets/rooms/rt2_s.tmx
	IGPE_Project/core/assets/rooms/rt3_e.tmx
	IGPE_Project/core/assets/rooms/rt4_n.tmx
	IGPE_Project/core/src/com/mygdx/game/Game.java
	IGPE_Project/core/src/com/mygdx/game/GameMain.java

+++ IGPE_Project/core/src/com/mygdx/game/model/weapons/ExplosionMagic.java
@@ -14,7 +14,7 @@ import com.mygdx.game.model.entities.Entity;
 //Konosuba reference
 public class ExplosionMagic extends Magic {
 
-	private int explosionDamage;
+	private float explosionDamage;
 	private float explosionRadius;
 	
 	public ExplosionMagic(Entity owner) {
